% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get}
\alias{data_get}
\title{get data from RIDL}
\usage{
data_get(ridldataset, dataraw = "data-raw")
}
\arguments{
\item{ridldataset}{name of the project in ridl.unhcr.org, the internal UNHCR repository based on CKAN}

\item{dataraw}{folder where to put the data}
}
\description{
get data from RIDL
}
\examples{

# data_get(ridldataset = "xxxx", dataraw = "data-raw")

}
