% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{get data from RIDL}
\usage{
data_load(dataraw = "data-raw")
}
\arguments{
\item{dataraw}{folder where to put the data}
}
\value{
progresobj  a list with all required frame
}
\description{
get data from RIDL
}
\examples{

#progresobj <- data_load(dataraw = "data-raw")

progresobj <- data_load(dataraw = system.file("demo/", package = "proGresAnalysis") )


}
