% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recode.R
\name{data_recode}
\alias{data_recode}
\title{Assemble all data components in a list and compute indicators}
\usage{
data_recode(progresobj, ctr)
}
\arguments{
\item{progresobj}{a list with all required frame}

\item{ctr}{Country or Business Unit for the dataset dump}
}
\value{
progresobjexpand  a list with all required frame
}
\description{
Assemble all data components in a list and compute indicators
}
\examples{
# progresobj <- data_load(dataraw = system.file("demo/", package = "proGresAnalysis") )
progres <- data_recode(progresobj = progresobj,
                       ctr = "Ecuador")

}
