% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reach_registration.R
\name{plot_reach_registration}
\alias{plot_reach_registration}
\title{Analytic chart to display registration coverage}
\usage{
plot_reach_registration(
  progres,
  ctr,
  ridl = "data-should-be-on-ridl",
  thisyear = 2021
)
}
\arguments{
\item{progres}{a list with all required frame re-encoded}

\item{ctr}{Country or Business Unit for the dataset dump}

\item{ridl}{name of the project in ridl.unhcr.org, the internal UNHCR repository based on CKAN}

\item{year}{Year of reference}
}
\value{
a list with both a ggplot2 crafted chart and the summary data
}
\description{
The function compute a summary chart
}
\examples{

# progresobj <- data_load(dataraw = system.file("demo/", package = "proGresAnalysis") )
# progres <- data_recode(progresobj = progresobj,  ctr = "Ecuador")
plot_reach_registration (progres = progres , 
                         ctr = "Ecuador", 
                         ridl = "Ecuador",
                                    thisyear = 2021)
}
