% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segment_var.R
\name{plot_segment_group}
\alias{plot_segment_group}
\title{Assemble all data components in a list and compute indicators}
\usage{
plot_segment_group(
  progres,
  ctr,
  ridl = "data-should-be-on-ridl",
  thisyear = 2021
)
}
\arguments{
\item{progres}{a list with all required frame re-encoded}

\item{ctr}{Country or Business Unit for the dataset dump}

\item{ridl}{name of the project in ridl.unhcr.org, the internal UNHCR repository based on CKAN}

\item{year}{Year of reference}
}
\value{
a list with both a ggplot2 crafted chart and the summary data
}
\description{
The function compute a summary chart with clusters relative description:
The variable modalities in each of the profiles can be __interpreted__ using the following information:  
  * _Cla/Mod_ is the pourcentage of records who have that modality and who are in that the profile (for instance: 50.3% of the cases who have a head of family who marrital status is single also belongs to profile 1) 
  * _Mod/Cla_ is the pourcentage of records in that profile who have that modality (for instance: 81.5% of the cases of profile  1 have a head of family who marrital status is single ) 
  * _Global_ is the global pourcentage of all records who have that modality (for instance: in all population, 51.67% of of the cases who have a head of family who marrital status is single.) 
  * _p-value_ provides an idea of the significant of the difference between Mod/Cla proportation & Global. If p-value is less than 0.05, then one category is significantly linked to another categories  
  * _v-test_ If the v-test is positive, it means that the category is over-expressed for the category and if the v-test is negative it means that the category is under-expressed for the category.What is interesting in the v-test is only the sign of the v-test. 
  
  Only variable modalities for which critical probability is below 0.02 are used."
}
\examples{

#progresobj <- data_load(dataraw = system.file("demo/", package = "proGresAnalysis") )
# progres <- data_recode(progresobj = progresobj,  ctr = "Ecuador")
# plot_segment_group(progres = progres,
#                                     ctr = "Ecuador",
#                                     ridl = "data-should-be-on-ridl",
#                                     thisyear = 2021 ) 
}
