% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segment_var.R
\name{plot_segment_var}
\alias{plot_segment_var}
\title{Assemble all data components in a list and compute indicators}
\usage{
plot_segment_var(
  progres,
  ctr,
  ridl = "data-should-be-on-ridl",
  thisyear = 2021
)
}
\arguments{
\item{progres}{a list with all required frame re-encoded}

\item{ctr}{Country or Business Unit for the dataset dump}

\item{ridl}{name of the project in ridl.unhcr.org, the internal UNHCR repository based on CKAN}

\item{year}{Year of reference}
}
\value{
a list with both a ggplot2 crafted chart and the summary data
}
\description{
The function compute a summary chart performing unsupervised classification on a series of key variables 
 https://www.datanovia.com/en/blog/cluster-analysis-in-r-simplified-and-enhanced/  
 https://www.datanovia.com/en/lessons/cluster-validation-statistics-must-know-methods/
 See video https://www.youtube.com/watch?v=0z30RDikYaw
}
\examples{

progresobj <- data_load(dataraw = system.file("demo/", package = "proGresAnalysis") )
progres <- data_recode(progresobj = progresobj,  ctr = "Ecuador")
cluster <- plot_segment_var(progres = progres,
                                    ctr = "Ecuador",
                                    ridl = "data-should-be-on-ridl",
                                    thisyear = 2021 )

cluster[[1]]

}
