% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_registration.R
\name{recode_registration}
\alias{recode_registration}
\title{Recode registration data from UNHCR proGres database and generate a summary report.}
\usage{
recode_registration()
}
\value{
save a cleaned csv file within the data folder.
}
\description{
Performs a series of cleaning
}
\examples{
\dontrun{
recode_registration()
}

}
\author{
Edouard Legoupil
}
